+++ src/SMAPI/Properties/AssemblyInfo.cs
@@ -1,5 +1,6 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("SMAPI.Tests")]
-[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")] // Moq for unit testing
+[assembly: InternalsVisibleTo("ConsoleCommands")]
 [assembly: InternalsVisibleTo("ContentPatcher")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")] // Moq for unit testing